# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AutoCrmApi",
    "version": "1.0"
  },
  "paths": {
...
    "/by-id/{id}": {
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}